+++ alf/algorithms/on_policy_algorithm.py
@@ -14,6 +14,7 @@
 """Base class for on-policy RL algorithms."""
 
 from abc import abstractmethod
+# from typing import Callable
 
 import tensorflow as tf
 
@@ -69,3 +70,14 @@ class OnPolicyAlgorithm(OffPolicyAlgorithm):
             observation=exp.observation,
             prev_action=exp.prev_action)
         return self.rollout(time_step, state)
+
+    # def prepare_specs(self,
+    #                   env_batch_size=None,
+    #                   time_step:ActionTimeStep=None,
+    #                   exp_replayer: str=None,
+    #                   metrics=[],
+    #                   observation_transformer: Callable=None):
+    def prepare_on_policy_specs(self, metrics=[]):
+        """ Prepare metrics.""" 
+
+        self._metrics = metrics
