+++ alf/algorithms/rl_algorithm.py
@@ -16,6 +16,8 @@
 from abc import abstractmethod
 from collections import namedtuple
 import itertools
+import os
+import psutil
 from typing import Callable
 
 import numpy as np
@@ -24,8 +26,10 @@ import tensorflow_probability as tfp
 
 from tf_agents.trajectories.time_step import StepType
 from tf_agents.utils import eager_utils
+from tf_agents.metrics import tf_metrics
 
 import alf.utils
+# from alf.utils import common, summary_utils
 
 TrainingInfo = namedtuple("TrainingInfo", [
     "action_distribution", "action", "step_type", "reward", "discount", "info",
@@ -113,6 +117,8 @@ class RLAlgorithm(tf.Module):
                  reward_shaping_fn: Callable = None,
                  train_step_counter=None,
                  debug_summaries=False,
+                 summarize_grads_and_vars=False,
+                 summarize_action_distributions=False,
                  name="RLAlgorithm"):
         """Create a RLAlgorithm.
 
@@ -153,6 +159,8 @@ class RLAlgorithm(tf.Module):
         self._predict_state_spec = predict_state_spec
         self._action_distribution_spec = action_distribution_spec
         self._optimizers = alf.utils.common.as_list(optimizer)
+        self._exp_observers = []
+        self._proc = psutil.Process(os.getpid())
         if get_trainable_variables_func is None:
             get_trainable_variables_func = lambda: super(RLAlgorithm, self
                                                          ).trainable_variables
@@ -170,6 +178,8 @@ class RLAlgorithm(tf.Module):
         self._train_step_counter = alf.utils.common.get_global_counter(
             train_step_counter)
         self._debug_summaries = debug_summaries
+        self._summarize_grads_and_vars = summarize_grads_and_vars
+        self._summarize_action_distributions = summarize_action_distributions
         self._cached_var_sets = None
         self._use_rollout_state = False
 
@@ -205,6 +215,20 @@ class RLAlgorithm(tf.Module):
     def trainable_variables(self):
         return list(itertools.chain(*self._get_cached_var_sets()))
 
+    @property
+    def exp_observers(self):
+        """Return experience observers."""
+        return self._exp_observers
+
+    # @abstractmethod
+    # def prepare_specs(self,
+    #                   env_batch_size,
+    #                   time_step:ActionTimeStep,
+    #                   exp_replayer: str,
+    #                   metrics,
+    #                   observation_transformer: Callable=None):
+    #     pass
+
     def _get_cached_var_sets(self):
         if self._cached_var_sets is None:
             # Cache it because trainable_variables is an expensive operation
@@ -219,6 +243,48 @@ class RLAlgorithm(tf.Module):
             step = self._train_step_counter
             tf.summary.histogram(name + "/value", rewards, step)
             tf.summary.scalar(name + "/mean", tf.reduce_mean(rewards), step)
+
+    def add_experience_observer(self, observer: Callable):
+        """Add an observer to receive experience.
+
+        Args:
+            observer (Callable): callable which accept Experience as argument.
+        """
+        self._exp_observers.append(observer)
+
+    def training_summary(self, training_info, loss_info, grads_and_vars):
+        if self._summarize_grads_and_vars:
+            alf.utils.summary_utils.add_variables_summaries(grads_and_vars,
+                                                  self._train_step_counter)
+            alf.utils.summary_utils.add_gradients_summaries(grads_and_vars,
+                                                  self._train_step_counter)
+        if self._debug_summaries:
+            alf.utils.common.add_action_summaries(training_info.action,
+                                        self._action_spec)
+            alf.utils.common.add_loss_summaries(loss_info)
+
+        if self._summarize_action_distributions:
+            alf.utils.summary_utils.summarize_action_dist(
+                training_info.action_distribution, self._action_spec)
+            if training_info.collect_action_distribution:
+                alf.utils.summary_utils.summarize_action_dist(
+                    action_distributions=training_info.
+                    collect_action_distribution,
+                    action_specs=self._action_spec,
+                    name="collect_action_dist")
+
+        for metric in self._metrics:
+            metric.tf_summaries(
+                train_step=self._train_step_counter,
+                step_metrics=self._metrics[:2])
+
+        mem = tf.py_function(
+            lambda: self._proc.memory_info().rss // 1e6, [],
+            tf.float32,
+            name='memory_usage')
+        if not tf.executing_eagerly():
+            mem.set_shape(())
+        tf.summary.scalar(name='memory_usage', data=mem)
 
     def greedy_predict(self, time_step: ActionTimeStep, state=None, eps=0.1):
         """Predict for one step of observation.
