+++ alf/drivers/policy_driver.py
@@ -16,7 +16,7 @@ import abc
 import os
 from typing import Callable
 
-import psutil
+# import psutil
 import math
 import gin.tf
 import tensorflow as tf
@@ -49,7 +49,7 @@ class PolicyDriver(driver.Driver):
 
         Args:
             env (TFEnvironment): A TFEnvoronmnet
-            algorithm (OnPolicyAlgorith): The algorithm for training
+            algorithm (RLAlgorithm): The algorithm for training
             observers (list[Callable]): An optional list of observers that are
                 updated after every step in the environment. Each observer is a
                 callable(time_step.Trajectory).
@@ -77,7 +77,7 @@ class PolicyDriver(driver.Driver):
             tf_metrics.AverageEpisodeLengthMetric(buffer_size=metric_buf_size),
         ]
         self._metrics = standard_metrics + metrics
-        self._exp_observers = []
+        # self._exp_observers = []
         self._use_rollout_state = use_rollout_state
 
         super(PolicyDriver, self).__init__(env, None,
@@ -92,54 +92,54 @@ class PolicyDriver(driver.Driver):
         self._observation_transformer = observation_transformer
         self._train_step_counter = common.get_global_counter(
             train_step_counter)
-        self._proc = psutil.Process(os.getpid())
+        # self._proc = psutil.Process(os.getpid())
         if training:
             self._policy_state_spec = algorithm.train_state_spec
         else:
             self._policy_state_spec = algorithm.predict_state_spec
         self._initial_state = self.get_initial_policy_state()
 
-    def add_experience_observer(self, observer: Callable):
-        """Add an observer to receive experience.
+    # def add_experience_observer(self, observer: Callable):
+    #     """Add an observer to receive experience.
 
-        Args:
-            observer (Callable): callable which accept Experience as argument.
-        """
-        self._exp_observers.append(observer)
+    #     Args:
+    #         observer (Callable): callable which accept Experience as argument.
+    #     """
+    #     self._exp_observers.append(observer)
 
-    def _training_summary(self, training_info, loss_info, grads_and_vars):
-        if self._summarize_grads_and_vars:
-            summary_utils.add_variables_summaries(grads_and_vars,
-                                                  self._train_step_counter)
-            summary_utils.add_gradients_summaries(grads_and_vars,
-                                                  self._train_step_counter)
-        if self._debug_summaries:
-            common.add_action_summaries(training_info.action,
-                                        self.env.action_spec())
-            common.add_loss_summaries(loss_info)
+    # def _training_summary(self, training_info, loss_info, grads_and_vars):
+    #     if self._summarize_grads_and_vars:
+    #         summary_utils.add_variables_summaries(grads_and_vars,
+    #                                               self._train_step_counter)
+    #         summary_utils.add_gradients_summaries(grads_and_vars,
+    #                                               self._train_step_counter)
+    #     if self._debug_summaries:
+    #         common.add_action_summaries(training_info.action,
+    #                                     self.env.action_spec())
+    #         common.add_loss_summaries(loss_info)
 
-        if self._summarize_action_distributions:
-            summary_utils.summarize_action_dist(
-                training_info.action_distribution, self.env.action_spec())
-            if training_info.collect_action_distribution:
-                summary_utils.summarize_action_dist(
-                    action_distributions=training_info.
-                    collect_action_distribution,
-                    action_specs=self.env.action_spec(),
-                    name="collect_action_dist")
+    #     if self._summarize_action_distributions:
+    #         summary_utils.summarize_action_dist(
+    #             training_info.action_distribution, self.env.action_spec())
+    #         if training_info.collect_action_distribution:
+    #             summary_utils.summarize_action_dist(
+    #                 action_distributions=training_info.
+    #                 collect_action_distribution,
+    #                 action_specs=self.env.action_spec(),
+    #                 name="collect_action_dist")
 
-        for metric in self.get_metrics():
-            metric.tf_summaries(
-                train_step=self._train_step_counter,
-                step_metrics=self.get_metrics()[:2])
+    #     for metric in self.get_metrics():
+    #         metric.tf_summaries(
+    #             train_step=self._train_step_counter,
+    #             step_metrics=self.get_metrics()[:2])
 
-        mem = tf.py_function(
-            lambda: self._proc.memory_info().rss // 1e6, [],
-            tf.float32,
-            name='memory_usage')
-        if not tf.executing_eagerly():
-            mem.set_shape(())
-        tf.summary.scalar(name='memory_usage', data=mem)
+    #     mem = tf.py_function(
+    #         lambda: self._proc.memory_info().rss // 1e6, [],
+    #         tf.float32,
+    #         name='memory_usage')
+    #     if not tf.executing_eagerly():
+    #         mem.set_shape(())
+    #     tf.summary.scalar(name='memory_usage', data=mem)
 
     @property
     def algorithm(self):
@@ -219,7 +219,7 @@ class PolicyDriver(driver.Driver):
                                    next_time_step)
             for observer in self._observers:
                 observer(traj)
-        if self._exp_observers:
+        if self._algorithm.exp_observers:
             action_distribution_param = common.get_distribution_params(
                 policy_step.action)
             exp = make_experience(
@@ -227,7 +227,7 @@ class PolicyDriver(driver.Driver):
                 policy_step._replace(action=action),
                 action_distribution=action_distribution_param,
                 state=policy_state if self._use_rollout_state else ())
-            for observer in self._exp_observers:
+            for observer in self._algorithm.exp_observers:
                 observer(exp)
 
         return next_time_step, policy_step, action
@@ -272,3 +272,27 @@ class PolicyDriver(driver.Driver):
     def _run(self, max_num_steps, time_step, policy_state):
         """Different drivers implement different runs."""
         raise NotImplementedError()
+
+    # def train(self,
+    #           experience: Experience,
+    #           num_updates=1,
+    #           mini_batch_size=None,
+    #           mini_batch_length=None):
+    #     """Train using `experience`.
+
+    #     Args:
+    #         experience (Experience): experience from replay_buffer. It is
+    #             assumed to be batch major.
+    #         num_updates (int): number of optimization steps
+    #         mini_batch_size (int): number of sequences for each minibatch
+    #         mini_batch_length (int): the length of the sequence for each
+    #             sample in the minibatch
+
+    #     Returns:
+    #         train_steps (int): the actual number of time steps that have been
+    #             trained (a step might be trained multiple times)
+    #     """
+    #     return self._algorithm.train(experience,
+    #             num_updates=num_updates,
+    #             mini_batch_size=mini_batch_size,
+    #             mini_batch_length=mini_batch_length)
